@@ -1,23 +1,32 @@
 {
-  "name": "boiler-template",
-  "version": "0.1.0",
+  "name": "toetsweb",
+  "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/jest": "^26.0.24",
-    "@types/node": "^12.20.21",
-    "@types/react": "^17.0.19",
-    "@types/react-dom": "^17.0.9",
-    "node-sass": "^6.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.2.1",
-    "react-scripts": "4.0.3",
-    "react-ts-boiler": "^1.1.5",
-    "typescript": "^4.4.2",
-    "web-vitals": "^1.1.2"
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.5.2",
+    "@types/node": "^16.18.3",
+    "@types/react": "^18.0.25",
+    "@types/react-dom": "^18.0.9",
+    "buro-lib-ts": "^1.0.26",
+    "chokidar": "^3.5.3",
+    "jszip": "^3.10.1",
+    "node-sass": "^7.0.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.2",
+    "react-scripts": "5.0.1",
+    "reactjs-popup": "^2.0.5",
+    "save-svg-as-png": "^1.4.17",
+    "typescript": "^4.9.3",
+    "web-vitals": "^2.1.4"
+  },
+  "devDependencies": {
+    "@types/react-beautiful-dnd": "^13.1.2",
+    "@types/react-dom": "^18.0.8",
+    "eslint-plugin-react": "^7.31.10"
   },
   "scripts": {
     "start": "react-scripts start",
