@@ -1,6 +1,6 @@
 enum Lang {
     NL = 'nl',
     EN ='en'
-};
+}
 
 export default Lang;
